Implement support for millisecond and microsecond times
This also deduces the time format rather than specifying it.  Time parsing
tries all the supported formats and saves the one that works to minimize
the try/except exception handling overhead.  The nice thing about this
is it makes it unnecessary to specify the time format when configuring
the TimeManager settings.

The current time setting in the project file gets truncated to the
previous second mark.  This is slightly annoying, but not annoying enough
for yours truly to fix.

Signed-off-by: Gerald Van Baren <gvb@unssw.com>