Don't duplicate original subset string when removing time restriction
The original subset string was stored locally and re-affected when the
time restriction was removed. Unfortunately the code used to set back
that original subset string was the same as the one used to set the
time restriction (`setSubsetString()`) and thus re-appended to itself.

This lead to huge duplicated SQL query that slowed down everything on
big data set (1 millions features).