Making it possible to have a zero based timeframe and still step
For WMS-time enabled it is often good to be able to step through
time slices not with a range like
start-end: TIME=2019-01-01T08:00/2019-01-01T12:00
but only with a 'moment' like
TIME=2019-01-01T08:00

To make this possible the commit makes it possible to add zero
as timeFrameLength again.
BUT(!) have a zero timeFrameLength would make it impossible to
go forward using the forward/back button. So this commit makes
it possible to use the timeFrameType dropdown (seconds, minutes...)
to step through time.
As an example:
http://geoservices.knmi.nl/cgi-bin/RADNL_OPER_R___25PCPRR_L3.cgi
will return a stack of images if you ask for a timeframe
But by using the zero-timefram in combi with for exampel 'week'
in the dropdown you can step through the data in weeks (or days,
or hours or ....)